requirements.txt
dbt-snowflake
apache-airflow
